Added authentication and role-based access control

+++ actions/delete_employee.php
@@ -1,5 +1,9 @@
 <?php
+session_start();
 include "../includes/db.php";
+if ($_SESSION['role'] != 'admin') {
+    die("Access denied");
+}
 $id = $_GET['id'];
 $conn->query("DELETE FROM employees WHERE id=$id");
 header("Location: ../index.php");
